Force PrepareSnapshotPartitionsForUpdate to fail in recovery
This forces update engine to fallback to overwriting existing partitions
legacy A/B style without creating any COW partitions.
This allows formatting data directly after flashing to be safely done,
without running the risk of wiping the update that was just flashed
on devices where there isn't enough space in super for COW devices.

It also solves the problem of having to be on slot B to be able to flash
when switching from stock to custom OS in the case of using a different
BOARD_SUPER_PARTITION_GROUPS compared to stock.

Conditionally allow creating snapshots for incrementals.

Change-Id: I89e80641a6154b5de4ecae6ccaf9b01cd73b0036
Signed-off-by: Anush02198 <Anush.4376@gmail.com>